api/requirements.txt
fastapi==0.115.11
uvicorn
pydantic
sqlalchemy
python-dotenv
psycopg2-binary